unquote values and fix root permissions

+++ setup.sh
@@ -1,5 +1,3 @@
-METRICBEAT_USER_ID=${METRICBEAT_USER_ID:-1000}
-
 METRICBEAT_DATA_DIR="/usr/share/metricbeat/data"
 if [[ "$EBS_VOLUME_NAME" != "" ]]; then
   if [[ "$EBS_VOLUME_DIR" != "" ]]; then
@@ -9,10 +7,12 @@ if [[ "$EBS_VOLUME_NAME" != "" ]]; then
   fi
 fi
 
-userdel metricbeat &>/dev/null
-adduser -u $METRICBEAT_USER_ID metricbeat -D
+if [[ "$METRICBEAT_USER_ID" != "" ]]; then
+  userdel metricbeat &>/dev/null
+  adduser -u $METRICBEAT_USER_ID metricbeat -D
 
-mkdir -p $METRICBEAT_DATA_DIR &>/dev/null
-chown -R metricbeat:metricbeat $METRICBEAT_DATA_DIR
+  mkdir -p $METRICBEAT_DATA_DIR &>/dev/null
+  chown -R metricbeat:metricbeat $METRICBEAT_DATA_DIR
 
-chown -R metricbeat:metricbeat /etc/rancher-conf/metricbeat
+  chown -R metricbeat:metricbeat /etc/rancher-conf/metricbeat
+fi
